Add GetProductEntrySerializer ProductEntry related get requests.
Because of the depth argument on ProductEntrySerializer, making
POST/PUT/PATCH/DELETE requests will required nested data as
as category object rather than the traditional PK. The new serializer
subclasses ProductEntrySerializer and the depth is now specified in the
new serializer. The ProductEntryList.get_serialzer_class is overwritten
to return GetProductEntrySerializer if request method is GET and
ProductEntrySerializer if otherwise.