fix(settings): Remove AWS_S3_OBJECT_PARAMETERS[CacheControl]
When posting images to galleria-bucket, this is triggered error instead:
ClientError at < -URL endpoint handling the upload- >
An error occurred (InternalError) when calling the PutObject operation (reached max retries: 4): We encountered an internal error. Please try again.

This is because CacheControl's max-age value is too high. By removing it, 'expires' is automatically assigned to '3600' by s3.